backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
torch==2.1.2
torchvision==0.16.2
numpy==1.24.3
opencv-python==4.9.0.80
Pillow==10.2.0
